Integrated upsimpler within the cross-validation setup, in the same way as within createModelsMinority.
Also included a check whether the upsimpler module is attached, as the cross-validation would seem to run with upsimpler but didn't in fact do so if the module was not attached.

Also implemented the 'default' upsampling setup within the majority classifier, where the lowly abundant tumor entities will not be removed anymore. This works, given the fact that it makes use of the kknn algorithm, which would only need one training sample to include it within the classifier creation.